+++ src/r3_core/src/bind.rs
@@ -172,7 +172,9 @@ use crate::{
     closure::Closure,
     hunk::Hunk,
     kernel::{self, cfg, prelude::*, raw, raw_cfg, StartupHook},
-    utils::{refcell::RefCell, ComptimeVec, ConstAllocator, Init, PhantomInvariant, ZeroInit},
+    utils::{
+        refcell::RefCell, ComptimeVec, ConstAllocator, Init, PhantomInvariant, ZeroInit, Zeroable,
+    },
 };
 
 mod sorter;
@@ -213,8 +215,10 @@ impl<T> BindData<T> {
     }
 }
 
+// FIXME: Derive this when <https://github.com/Lokathor/bytemuck/pull/148> is
+//        merged
 // Safety: Zero-initialization is valid for `MaybeUninit`
-unsafe impl<T> ZeroInit for BindData<T> {}
+unsafe impl<T> Zeroable for BindData<T> {}
 
 // Main configuration interface
 // ----------------------------------------------------------------------------